requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
pandas>=1.5.0
numpy>=1.24.0
customtkinter>=5.2.0
pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
requests>=2.31.0